Work in Progress, Updating LabClass to include course
Created Course, with the imputable variables of Title and Code, Still looking into how to assign the course to the student, only if the student is enrolled in the class. Also looking into adding it into the print out, but may not be able untill i've figure out how to add the course to the enrolled students.